backend/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.42.0
django-cors-headers==3.5.0
django-extensions==3.0.8
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.10
joblib==0.14.0
JPype1==1.0.2
konlpy==0.5.2
lxml==4.5.2
numpy==1.19.2
oauthlib==3.1.0
pandas==0.25.3
Pillow==7.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sqlparse==0.3.0
threadpoolctl==2.1.0
tweepy==3.9.0
urllib3==1.25.10
